+++ web/client/js/fetch.js
@@ -21,40 +21,47 @@ async function getUserDatas(token) {
         else return false
     })
 }
-
-async function getUser(token) {
+async function getAccount(token) {
     if(!token) return false
-    const {type, access} = token
-    if(!access || !type) return false
-
-    return fetch('https://discord.com/api/v9/users/@me', {
-        headers: {
-            authorization: `${type} ${access}`,
-            "Content-Type": "application/json"
-        }
-    }).then(res => {
+    
+    const {value} = await fetch('/client?token=' + JSON.stringify(token)).then(res => {
         if(res.ok) return res.json()
         else return false
     })
+    if(!value || !value.user || !value.guilds) return undefined
+    let {user, guilds} = value
+
+    guilds = await fetch('/permissions', {
+        body: JSON.stringify(guilds), 
+        method: "POST"
+    }).then(r => r.json())
+    .then(res => {
+        if(res.code === 200) {
+            return res.value
+        }else return false
+    })
+
+    return {user, guilds}
+}
+async function getUser(token) {
+    if(!token) return false
+
+    return fetch('/client?token=' + JSON.stringify(token)).then(res => {
+        if(res.ok) return res.json()
+        else return false
+    }).then(elements => elements.value?.user)
 }
 async function getGuilds(token) {
     if(!token) return false
-    const {type, access} = token
-    if(!access || !type) return false
 
-    return fetch('https://discord.com/api/v9/users/@me/guilds', {
-        headers: {
-            authorization: `${type} ${access}`,
-            "Content-Type": "application/json"
-        }
-    }).then(res => {
+    return fetch('/client?token=' + JSON.stringify(token)).then(res => {
         if(res.ok) return res.json()
         else return false
     })
     .then(async srv => {
-        if(!srv) return false
+        if(!srv || !srv.value?.guilds) return false
         const filtered = await fetch('/permissions', {
-            body: JSON.stringify(srv), 
+            body: JSON.stringify(srv.value.guilds), 
             method: "POST"
         }).then(r => r.json())
         .then(res => {