+++ web/client/js/guildsDatas.js
@@ -1,9 +1,0 @@
-async function getGuildsDatas(token) {
-    const guilds = await getGuilds(token)
-    const guild_card = await getTemplate('guild_card')
-    
-    return {
-        guilds,
-        guild_card
-    }
-}